accessible/tests/mochitest/test_states_frames.html

<html>

<head>
  <title>frame based document testing</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/states.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      frameDoc = document.getElementById("frame_doc").contentDocument;
      frameDocArticle = document.getElementById("frame_doc_article").contentDocument;
      frameDocCheckbox = document.getElementById("frame_doc_checkbox").contentDocument;
      frameDocTextbox = document.getElementById("frame_doc_textbox").contentDocument;
      
      testStates(frameDoc, STATE_READONLY);
      testStates(frameDocArticle, STATE_READONLY);
      testStates(frameDocCheckbox, 0, 0, STATE_READONLY);
      testStates(frameDocTextbox, 0, EXT_STATE_EDITABLE);

      frameDoc.designMode = "on";
      testStates(frameDoc,  0, EXT_STATE_EDITABLE);
      testStates(frameDocArticle, STATE_READONLY);
      testStates(frameDocCheckbox, 0, 0, STATE_READONLY);
      testStates(frameDocTextbox,  0, EXT_STATE_EDITABLE);

      frameDocArticle.designMode = "on";
      testStates(frameDocArticle, 0, EXT_STATE_EDITABLE);

      frameDocCheckbox.designMode = "on";
      testStates(frameDocCheckbox, 0, 0, STATE_READONLY);
      
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>

<body>

  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=467387"
    title="Expose non-editable documents as readonly, regardless of role">
     Mozilla Bug 467387
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
  
  <iframe id="frame_doc" src="z_states_frame.html"></iframe>
  <iframe id="frame_doc_article" src="z_states_framearticle.html"></iframe>
  <iframe id="frame_doc_checkbox" src="z_states_framecheckbox.html"></iframe>
  <iframe id="frame_doc_textbox" src="z_states_frametextbox.html"></iframe>
</body>
</html>
